chg: refactoring of additional plot style controls fix: ColorComboBox and BrushComboBox now show also ColorSetLevel in first line fix: BrushComboBox now show reasonable brush names fix: EventSuppressor bug causing exception

+++ Altaxo/Base/Graph/Gdi/Plot/Styles/ScatterPlotStyle.cs
@@ -466,12 +466,12 @@ namespace Altaxo.Graph.Gdi.Plot.Styles
 			CreateEventChain();
 		}
 
-		public ScatterPlotStyle(XYPlotScatterStyles.Shape shape, XYPlotScatterStyles.Style style, float size, float penWidth, NamedColor penColor)
+		public ScatterPlotStyle(XYPlotScatterStyles.Shape shape, XYPlotScatterStyles.Style style, double size, double penWidth, NamedColor penColor)
 		{
 			_shape = shape;
 			_style = style;
 			_dropLine = new CSPlaneIDList();
-			_pen = new PenX(penColor, penWidth);
+			_pen = new PenX(penColor, (float)penWidth);
 			_symbolSize = size;
 
 			_relativePenWidth = penWidth / size;
